Package code
Organize the classes into suitable java packages.
