+++ src/main/java/duke/Storage.java
@@ -1,3 +1,9 @@
+package duke;
+
+import duke.exception.CorruptedFileException;
+import duke.exception.DukeException;
+import duke.task.Task;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
