+++ src/main/java/duke/command/DukeCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.DukeList;
+import duke.exception.DukeException;
+
 public class DukeCommand {
 
     public DukeCommand() {}
