+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public class ExitCommand extends DukeCommand {
 
     public ExitCommand() {
