+++ src/main/java/duke/exception/CorruptedFileException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class CorruptedFileException extends DukeException{
     public CorruptedFileException() {
         super("Corrupted File");
