+++ src/main/java/duke/exception/InvalidArgumentsException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class InvalidArgumentsException extends DukeException{
     public InvalidArgumentsException() {
         super("Invalid or Missing Arguments");
