+++ src/main/java/duke/exception/InvalidTaskException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class InvalidTaskException extends DukeException{
     public InvalidTaskException() {
         super("Invalid Command");
